server/requirements.txt
fastapi[standard]==0.115.4
motor==3.6.0
passlib==1.7.4
PyJWT==2.9.0
prettytable==3.12.0